+++ tests/test_cli_help.py
@@ -17,14 +17,13 @@ class TestExample:
             .reserve("lifespan", dtype=Period)
         ).build()
         z = DfCliHelp.help(clazz)
-        txt = z.get_full_text()
+        txt = z.get_long_text()
         assert "pretty bird table" in txt
         assert "table of species" in txt
         assert "- species (string)" in txt
         assert "- prettiness (floating-point)" in txt
         assert "- cuteness (integer)" in txt
         assert "- lifespan (time period)" in txt
-        assert "- .csv[.zip/.gz/.bz2/.xz]: Comma-delimited" in txt
         assert "[discouraged]" in txt
         assert "\n\n" in txt
 
@@ -35,7 +34,7 @@ class TestExample:
             .dtype(bool)
         ).build()
         z = DfCliHelp.help(clazz)
-        txt = z.get_full_text()
+        txt = z.get_long_text()
         assert "whether they like each other" in txt
         assert "cast to bool" in txt